feat: restructure code and tests
