+++ Daf.Oche.sln
@@ -5,15 +5,15 @@ VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{278304AF-2984-476E-8C25-A957C87D01E5}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Daf.Oche.Cli", "src\Daf.Oche.Console\Daf.Oche.Cli.csproj", "{BB9EBD35-132A-4A6E-A675-86D7139C07DB}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{F33AECF4-1519-4B6A-BDD3-3515022DD24A}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Daf.Oche.Cli.Tests", "tests\Daf.Oche.Console.Tests\Daf.Oche.Cli.Tests.csproj", "{6E1B801D-9E83-4407-8F90-452555D43942}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Daf.Oche.Domain", "src\Daf.Oche.Domain\Daf.Oche.Domain.csproj", "{F3A196AF-0489-46BC-8B2D-026B0125DEEC}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Daf.Oche.Domain.Tests", "tests\Daf.Oche.Domain.Tests\Daf.Oche.Domain.Tests.csproj", "{9733DEFA-A468-4FCA-BCE0-55F7D5CC6D3D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Daf.Oche.Cli", "src\Daf.Oche.Cli\Daf.Oche.Cli.csproj", "{EA60B979-D10A-40DC-BFAE-5A48FAEDD1A8}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Daf.Oche.Cli.Tests", "tests\Daf.Oche.Cli.Tests\Daf.Oche.Cli.Tests.csproj", "{2D31D37E-8088-4117-AAC5-057471903CDC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,14 +21,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{BB9EBD35-132A-4A6E-A675-86D7139C07DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BB9EBD35-132A-4A6E-A675-86D7139C07DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BB9EBD35-132A-4A6E-A675-86D7139C07DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BB9EBD35-132A-4A6E-A675-86D7139C07DB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6E1B801D-9E83-4407-8F90-452555D43942}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6E1B801D-9E83-4407-8F90-452555D43942}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6E1B801D-9E83-4407-8F90-452555D43942}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6E1B801D-9E83-4407-8F90-452555D43942}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F3A196AF-0489-46BC-8B2D-026B0125DEEC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F3A196AF-0489-46BC-8B2D-026B0125DEEC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F3A196AF-0489-46BC-8B2D-026B0125DEEC}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -37,15 +29,23 @@ Global
 		{9733DEFA-A468-4FCA-BCE0-55F7D5CC6D3D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9733DEFA-A468-4FCA-BCE0-55F7D5CC6D3D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9733DEFA-A468-4FCA-BCE0-55F7D5CC6D3D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EA60B979-D10A-40DC-BFAE-5A48FAEDD1A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EA60B979-D10A-40DC-BFAE-5A48FAEDD1A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EA60B979-D10A-40DC-BFAE-5A48FAEDD1A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EA60B979-D10A-40DC-BFAE-5A48FAEDD1A8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2D31D37E-8088-4117-AAC5-057471903CDC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2D31D37E-8088-4117-AAC5-057471903CDC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2D31D37E-8088-4117-AAC5-057471903CDC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2D31D37E-8088-4117-AAC5-057471903CDC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{BB9EBD35-132A-4A6E-A675-86D7139C07DB} = {278304AF-2984-476E-8C25-A957C87D01E5}
-		{6E1B801D-9E83-4407-8F90-452555D43942} = {F33AECF4-1519-4B6A-BDD3-3515022DD24A}
 		{F3A196AF-0489-46BC-8B2D-026B0125DEEC} = {278304AF-2984-476E-8C25-A957C87D01E5}
 		{9733DEFA-A468-4FCA-BCE0-55F7D5CC6D3D} = {F33AECF4-1519-4B6A-BDD3-3515022DD24A}
+		{EA60B979-D10A-40DC-BFAE-5A48FAEDD1A8} = {278304AF-2984-476E-8C25-A957C87D01E5}
+		{2D31D37E-8088-4117-AAC5-057471903CDC} = {F33AECF4-1519-4B6A-BDD3-3515022DD24A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {41C7210E-4D3F-4896-84CC-581CFDF4F601}
