+++ src/Daf.Oche.Domain/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Daf.Oche.Domain.Tests")]
+[assembly: InternalsVisibleTo("Daf.Oche.Cli.Tests")]
