+++ tests/Daf.Oche.Cli.Tests/Daf.Oche.Cli.Tests.csproj
@@ -13,16 +13,18 @@
     <PackageReference Include="coverlet.collector" Version="6.0.0" />
     <PackageReference Include="FluentAssertions" Version="6.12.2" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0" />
+    <PackageReference Include="Moq" Version="4.20.72" />
     <PackageReference Include="xunit" Version="2.5.3" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.5.3" />
   </ItemGroup>
 
   <ItemGroup>
-    <Using Include="Xunit" />
+    <ProjectReference Include="..\..\src\Daf.Oche.Cli\Daf.Oche.Cli.csproj" />
+    <ProjectReference Include="..\..\src\Daf.Oche.Domain\Daf.Oche.Domain.csproj" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Daf.Oche.Console\Daf.Oche.Cli.csproj" />
+    <Using Include="Xunit" />
   </ItemGroup>
 
 </Project>
